Implementado sistema de mensageria no sistema, pedidos sendo realizados com sucesso

+++ src/main/java/com/site/toffCo/module/produto/dto/ProdutoRequestDTO.java
@@ -2,15 +2,13 @@ package com.site.toffCo.module.produto.dto;
 
 import jakarta.validation.constraints.NotBlank;
 
-import java.math.BigDecimal;
-
 public record ProdutoRequestDTO(
         @NotBlank(message = "Nome is obrigatory")
         String name,
         @NotBlank(message = "Description")
         String description,
         @NotBlank
-        BigDecimal price,
+        Float price,
         String categoria,
         String image,
         Integer estoque
